Add site.yml, network role & localhost config
First and foremost, this adds the site.yml, which really should have
been done earlier. Now this can actually be used and tested.

This also adds the network role, which for now only handles basic
hostname setup. However, because the goal is to be able to use this as
both a local bootstrap and for remote management/support, some things
had to be considered.

A localhost.yml file is introduced to solve the need for initally
setting the hostname. This is done via the "arch_hostname: hostname"
key/value pair.

The idea is that if you later use another inventory (i.e for remote
management) this value is no longer required, because the
{{ arch_hostname }} variable will default to {{ ansible_hostname }},
which will then be available.