Adds root role, for setting root password
This commit introduces the root role, that will be used to set the root
password. Do note however, that this is optional. The root_pass and
root_pass_salt variables can be left undefined (commented out), which
will make ansible skip the task.

Security Notes

This may be a small role, but it is extremly important to get right.
Regardless of the security in the choices made, the complete security
implications of this commit should probably be revisited and reviewed
in the future. Perhaps by someone more experienced in such matters.

There are some conflicting documentation about passwords in the
Ansible documentation. They warn against using plaintext passwords in
your host_vars, yet recommend the Ansible Vault, which could be used
on any of these files.

In the end it was decided this was the best solution at this point in
time.

Some alternative discarded ideas:

 - Using vars_promt. These are run every time, and don't scale well
   since you might need unique root passwords on a per host basis.
 - Using environment variables. These should never be trusted with
   secure data and they too don't scale well.
 - Using a separate vault encrypted file. This does not allow setting
   on a per host basis.