Adds reflector role, for pacman mirror updates
Reflector is a script and package for Arch Linux which manages your
pacman mirrorlist, keeping it up to date and fast. Read more at:
https://wiki.archlinux.org/index.php/Reflector

The included systemd units and pacman hook makes this an idea automated
way of managing mirrors that otherwise often get stale for various
reasons. Even if the user knows how to update them they may opt not to,
perhaps because of laziness or because "it just works". A user would
also need to know when new mirrors are added or old ones removed, which
is hard if you don't make a routine out of it. This commit fixes this
by adding support such routines in an automated manner.

Since this project already pretty much assumes a Swedish locale, this
new role will limit mirrors to ones located in Scandinavia. At least
for now.