Adds users role, for users & ssh-key management
This new role adds initial support for user management. Its
configuration is heavily inspired by the format of cloud-init, but there
are some obvious differences, such as the use of underscore wherever
dash would be used in variable names. This is due conflicts in how
Ansible interprets variables.

See cloud-init documentation at:
https://cloudinit.readthedocs.io/en/latest/index.html

Other differences include:
- "passwd:" is currently the equivalent of "plain_text_passwd:"
  Although this will likely be changed when hashed password support is
  added.
- "passwd_salt:" is an addition that adds a user defined salt to your
   plain text password.
- "state:" is a simple way of defining if a user should exist or be
  deleted on next run.
- "username:" is the equivalent of "name:"
- "name:" is a more user friendly name for "gecos:", which is more
  relvevant here since this project is initially aimed at desktop systems.

A great deal of effort has been put in to make sure nothing (except
"username:", duh..) is mandatory. Anything not defined by the user will
be omited or in a minority of cases a default value will be used (such
as the shell).