Adds sudo setup role
This role could have been part of users. However that would have made
the users role big since sudo is big in itself. Some may also prefer not
to use sudo at all, so this is more modular.

For now sudo access is granted with the use of a single group. At this
time there is little motivation for allowing multiple groups, although
this could technically be added without much difficulty.

There is also no support for allowing sudo on a per user basis, although
this is on the TODO list. This is not a big priority since it is easy to
add the user to a group like wheel and use "sudo_group: wheel". The
wheel group even exist by default.

If per user sudo support is added it will likely reuse the "users" list,
also used by the users role. This way we can keep replicating the
cloud-init format, where appropriate.