Adds package upgrade & multilib support
The package role will now check for package upgrades before it installs
user defined packages.

Multilib support is enabled via a custom pacman config file, which is
included in pacman.conf.

The reason for doing it this way is the square brackets in pacman.conf,
which causes problems with ansible and the lineinfile module. We could
also solve this with a custom pacman.conf file, but since Arch Linux
uses rolling release upgrades, this would mean constant catch up work.