Fixes and simplifies pacman multilib support
It turns out the pacman.conf format was incorrect and still commited. A
simple Include line in the bottom of the file, and nothing else is
enough. Sections are then defined in the new config file.

Files and mentions are now renamed to "multilib" and any assumption of
the file being used for other things (in the future) is removed. If this
changes in the future it is simple to change back.