Adds a very specialized yubikey role
This is a big role. It could easily have been divided into a gnupg,
gpg-agent-ssh (ssh-agent) and possibly an opensc/pcsc role. However,
the specifics of this role heavily rely upon each other and are very
error prone if not done carefully. At this time this is the only way
I (the author) use ssh and pgp private keys so ultimately it makes
sense.

This role makes sure that all current and future users have out of the
box support for Yubikeys. This includes using U2F, OpenPGP
(for SSH, among other things) and also pure RSA keys (for SSH).

Currently these settings will be forced upon all users and most of it is
not possible to override. If this presents issues in the future, at the
very least, a per user option could perhaps be added.

Sources of information that this role is based upon:
- https://wiki.archlinux.org/index.php/GnuPG#SSH_agent
- https://wiki.archlinux.org/index.php/GNOME/Keyring#Disable_keyring_daemon_components