Adds configurable primary group to users role
There are some big questions regarding expectations as to what will be
done when "primary_group" is defined but said group does not exist. This
mighty vary depending on if other group management solutions are in use,
where conflicts might occur, or not.

Therefore this behaviour is made configurable with
"users_create_primary_group", which will take a boolean value. The
default is true for two reasons.

1. User friendlieness, or simple expectations of things just working.
2. If "primary_group" is set and "users_create_primary_group" is false it
   will fail, unless groups are specifically created elsewhere.

A primary concern here is mandatory external dependencies for internal
features. A role should in most cases work by itself with all features.

Additional minor fixes:
- Not defining "passwd_salt" when setting a password results in failure.
- Logging is now disabled for all tasks, since they all use the "users"
  list. This otherwise exposes user passwords.