Adds a basic hidpi scaling role
This role attempts to automate as much as possible from the Arch Wiki
page: https://wiki.archlinux.org/index.php/HiDPI

Desktop Environments is currently limited to Gnome/GTK based ones,
because that is the focus of this project. Initial support for
applications with special separate settings is also included. At this
time this only includes Spotify, but Steam (and maybe others) will be
added.