Add support for Steam in hidpi role
Steam is now supported by using the Steam-HiDPI-Skin, available on
GitHub: https://github.com/MoriTanosuke/HiDPI-Steam-Skin

This role does rely on git now and this is not supplied. A package
installation task might have to be added to ensure this never fails.

A hard coded assumption that a user's primary group uses the same name
is also removed now. Instead tasks rely on the become_user mechanism.