Add modular boot manager support
Alternative boot loaders are now available with conditional roles and
a variable "boot_manager".

The new bootloader added is grub. This will be used by default if the
machine is booted in legacy/non-uefi mode. For now this is limited
to 32-bit installations.

If UEFI mode is detected systemd-boot will be used by default.

A lot of conditional complexities has been moved to the main playbook,
instead of doing it on the task level for each role.