Handle efi boot variable entries in systemd-boot
The systemd-boot role will now add an entry to the EFI boot variables
and add it to the top of the boot order. This was not added until now
because it was believed that without proper conditionals an infinate
amount of entries would be added, filling the NVRAM. It would appear
this is not the case.