Add a fstab role for managing mounts
The primary goal for this role is not to manage all your mounts. In a
typical Arch Linux installation this is already done by the pacstrap
script. Besides that, conflicting user customizations are a too likely
scenario.

Instread it will appy minimal modifications including:
- Make sure /tmp is using tmpfs
- Make the efivarfs mounted read-only (UEFI boot mode only)
- Apply the mount options "nodev,nosuid,noexec,noatime" to the above

After this /etc/fstab will be formated (columnated) for easy reading.
For added safety this work is validated before applied.

Because this is a sensitive area the role can easily be disabled with
"fstab.enabled: false". This is set to true by default.

If in the future operators do wish to use this for managing all fstab
entries, that could easily be added.