Add the foundation for a aur wrapper role
The idea with this role is to automatically build and install AUR
packages. There are some serious security implications with this, so
minimal use is recommended.

Primarily this is good for an initial bootstrap of AUR helper tools. Do
note that at this time there is no simple way to install yaourt, because
that requires more than one package (package-query). The goal is to
solve this.