Fix issues with ansible_managed files
When the ansible_managed variable was added to the head of files
the use of the copy module should have been changed to the template
module. This commit rectifies the error of missing this.

Also fixes some issues with octal notations when being passed as
variables, which would sometimes lead to incorrect permissions.