Rework of systemd-boot role
This introduces some major changes to various parts of the
systemd-boot role. The biggest is probably the remounting tasks, that
will make sure the efivarfs is mounted with read-write access. This
is added since this task would otherwise no longer function since the
addition of the fstab role.

One thing to keep in mind is if the efivarfs filesystem is not mounted
this role will now fail. However, at this time this is acceptable,
since this is already used as a conditon for the role itself.

Other changes include:
- The esp mountpoint is now configurable but defaults to the
  path /boot, which is recommended by the Arch Wiki.
- A pacman hook is added to ensure systemd updates also updates the
  efi executables in the esp.
- File and directory tasks are now unified under two tasks.
- Octal notations passed through variables are now qouted because of
  a strange issue with incorrect permissions.