Fix network role failing in chroot
The failure is caused by the hostname module which uses the
hostnamectl method by default on ArchLinux. Because the module does
not handle failures and fallback to another method this has to be
handled at a task level.

This commit re-adds the previous way of handling hostnames and falls
back to this on failures. It also skips the failure part entirely if
chroot_mode is set. Later will be a true ansible fact and renamed to
ansible_chroot.