Fix pacman role order failing other roles
Since the pacman role was moved it would cause failures on roles that
required multilib configured. This moves it back up in the order.

The reasoning originally behind the move was that personal package
installation belongs in the end. But clearly the rest belongs at an
earlier stage. How to best solve this?

1. Accept it like this?
2. Split user defined package installation into a new role?
3. Have a handler for it? But this would break the yaourtrc task in
   the aur role...