Add filesystem dependency handling
Filesystems like f2fs requires f2fs-tools to be installed, or the
neccessary fsck tool will be absent. This would cause the mkinitcpio
handler to fail and your system might fail to boot. This is and
other dependencies are now handled in a reusable and easily
extended fashion.

Minor related fixes/changes:
- grub role updated to fix mkinitcpio handler conflict
- fstab tag added to some tasks so when, for example new filesystems
  are added, the mkinitcpio role will run if neccessary