Add menu role for hiding desktop menu entries
This adds the capability to hide some annoying desktop entries that
are common on Arch Linux systems.

The root of the issue is that the distribution package policy means
packages are based on unmodified upstream sources. This is generally
a good thing, but it also means packages are rarely split into cli
and gui verisons. So simply installing base-devel for aur, adds a lot
of desktop entries irrelevant to most users.

So to solve this a new role is added that includes an optional preset
list of the most common desktop entries. In addition, a custom user
defined list is also supported.

If anything the only caveat at this time is the inability to change
an entry back from hidden (NoDisplay=true) to visable again. This
should be easy to add though if neccessary.