Add configurable locale
The locale role now accepts configurable system loclales (LANG) and
vc keymaps. In addition to the value oflocale_lang, en_US.UTF-8 is
generated by default. This is because various applications expect
the english language. If not required it is easily overridden with
the locale_gen variable.

X11 layout is not managed at this time; still considering if this
belongs in the xorg role.