Remove sudo role and replace with users role
Everything related to sudo is now integrated into the users role, so
the old sudo role is not needed anymore.

There were several reasons for this, one of which was the
impossibility of removing all sudo enabled users and then adding new
ones. This would fail since when the sudo role ran it would have cut
itself off.

It is possible this is still the case, but now it should be possible
to fix, for example by splitting off user removal into its own task
at the very end. Although it would also be required to change
ansible_user on the fly, is this possible?

It also seemed the sudo role served no purpose by default. It would
just duplicate what was already there. What actually gave sudo was
adding a user to a group in the users role.

Therefore the new sudo functionality is user based instead. If a
group based solution is wanted, simply add your user to the wheel
group (or sudo on Debian, or sudo/admin on Ubuntu).