Add custom default dconf values
The gnome role will now customize the default dconf database values.
Most importantly, this adds the ability to set the keyboard layout
for the graphical environment. It also adds some nice things like
ctrl+alt+t to open the terminal, disabling the system bell, a nice
preset exclude list for deja-dup and much more.

This could be made optional in the future, but at this time it is
not possible. These are default values only, easily overriden.