Refactor and simplify menu role
Removes all tasks related to reusing files existing on hosts. This
was done for various reasons. Some of them being:

- Performance was quite bad.
- Breaks idempotence when tmpfs is used for menu_tmp_path (default).
- The task got a lot more complicated than felt neccessary.

The decider was the fact that secure and reliable temporary storage
is hard. XDG_RUNTIME_DIR turns out to be a bad choice, because
whenever the root user is used this is unset resulting in /ansible
being created.

One could argue that these files are not that security critical, but
to me /tmp remains a bad alternative. Besides, that would not solve
idempotence issues and would also require an added tempfile task
(which lacks a module pre 2.3).