Update grub role
It is now possible to do user defined boot options, using the
boot_options variable. Mind, that it will not check for conflicts
with any existing options, so don't use "boot_options: root=..."
simultanious with boot_encryption.

Regarding that, according to the arch wiki grub should be able to
get root device automatically. If disk encryption is only planned
to be supported as it is now then rootdevice should not be needed.
So perhaps this should be explored next.