Replace lookups with register tasks
All current lookups were used in a way that assumed the hosts were
local. This has worked so far, because all testing has been done on
localhost, which also was the main use case (live install).

This should help in getting the project to work on remote hosts.
However testing remains on the TODO list.