Work around ansible 2.2.1 service module issue
Ansible point release 2.2.1 breaks enabling services in a chroot on
systems using systemd init.

Previously chroot_mode was implemented to exempt starting services in
a chroot. With this new issue the service, or systemd module
litterally cannot do anything useful in a chroot environment.

Therefore all service enable/start tasks will now use the block and
error handling features introduced in Ansible 2.0. By default the
services are enabled and started normally. If a failure is caught,
the play falls back to enabling units with symlinks.

The end result is everything now works with Ansible 2.2.1, services
are always enabled, and any future issues with the service modules
should be handled.