Fix nvidia libgl issues
Whenever the nvidia role was used tasks would fail because another
libgl provider is installed. Ideally we want to support both and
when mesa + libglvnd support is ready it won't be an issue.

The cause of the issues seems to be changes in how Arch Linux handle
libgl. By default mesa-libgl is used, which means installing nvidia
now won't use the correct libgl package.

The following solved this:
- Remove the xorg role, since xorg-server has libgl as a dependency.
- Trick the pacman module to install nvidia,nvidia-libgl
  (and lib32-nvidia-libgl if x86_64) in the same command, instead of
  one by one which would install mesa-libgl.

Additionally, because everything is now handled in a single pacman
command, the nvidia role is now much better optimized.