Replace locale, time-zone & hostname roles
This introduces a reusable sys-config role that replaces the mentioned
roles. The plan is to make this fully distribution agnostic.

TODO: A side effect of this is the gnome role is now in a bad position
because a dconf override uses sys_config_locale_lang. This should be
replaced with some alternative dynamic way of getting that string.