Enable using dictionary/hashes with gsettings
The new task uses with_dict to support a more compact set of hashes.
It does not support different users so it remains complimentary to
the existing input formats.

The gsettings role is updated to use blocks, explicit become:
true/false and will now report skipped in favor of being silent.