Replace gsetting module with rewrite
The reasons behind this were plentyful:
- No more uncertainties about the license.
- The possibility of 100% BSD licensed code.
- More than 90% of the code was already rewritten anyhow.
- It offered a clean start and plenty of learning opportunities.

The new module will prefer the cleaner dbus-run-session command.
If this command cannot be found (dbus version < 1.8.0, I believe),
it will fallback to the older dbus-launch command and kill leftover
processes. The best of both worlds, and it is easy to deprecate
the legacy support in the future.

It is hosted on GitHub at:
https://github.com/AsavarTzeth/ansible-gsetting