Fix syntax warning in fstab role
Unfortunately fstab role depended on this invalid syntax to properly
skip a task. So that had to be fixed as well.

A positive side effect is that the role will now also ensure that there
are no left over efivars entries in /etc/fstab. This is useful if, for
example, a system has been converted or migrated away from UEFI.